+++ app/Http/Controllers/Admin/DashboardController.php
